+++ Sources/rock/main.swift
@@ -31,6 +31,21 @@ main.register(UninstallCommand())
 main.register(InitCommand())
 main.register(VersionCommand())
 main.register(RunCommand())
+main.register(ScriptCommand(
+  verb: "build",
+  defaultScript: RockConfig.rockConfig.debugBuildScript,
+  function: "Executes the script defined in Rockfile named build."
+))
+main.register(ScriptCommand(
+  verb: "test",
+  defaultScript: RockConfig.rockConfig.testScript,
+  function: "Executes the script defined in Rockfile named test."
+))
+main.register(ScriptCommand(
+  verb: "archive",
+  defaultScript: RockConfig.rockConfig.archiveScript,
+  function: "Executes the script defined in Rockfile named archive."
+))
 main.register(HelpCommand(registry: main))
 
 main.main(defaultVerb: "help") { error in