+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - **[Rock]** Now supports `rock version` - @vknabel
 - **[Rock]** Now supports `rock run script` - @vknabel
+- **[Rock]** Added convenience scripts for `rock build`, `rock test` and `rock archive` - @vknabel
+- **[Rock]** Derived scripts that will be executed will be highlighted now - @vknabel
 - **[Rockfile]** Every value in `Rockfile` will now be interpreted as Stencil template - @vknabel
 - **[Rockfile]** Overriding single properties in `Rockfile` is now supported - @vknabel
 - **[Other]** Provides experimental `Install.generated.swift` Installation Script - @vknabel